fix(styles): pagination component update due to new designs (#2973)
BREAKING CHANGE:
If > 9 pages then 9 elements should be shown - including page links, more, current page input. Otherwise, all pages should be shown, use the link instead of input with .is-active class or aria-selected="true" attribute for the current page. Input should still be present to be shown in the mobile mode, but it's hidden in desktop mode if < 9 pages.

Add an additional label with class .fd-pagination__label after the input in the mobile mode with the number of pages available.